fix: use temp dir instead of overwrite
