+++ .github/workflows/deploy-package.yml
@@ -9,7 +9,7 @@ on:
         required: true
 
 jobs:
-  publish:
+  publish-sdk:
     runs-on: ubuntu-latest
 
     steps:
@@ -71,38 +71,33 @@ jobs:
         working-directory: packages/sdk
         run: npm publish --access public
 
-      # Build and publish wagmi connector package
-      - name: Backup original package.json and README
-        working-directory: packages/sdk
-        run: |
-          cp package.json package.json.backup
-          cp README.md README.md.backup
+  publish-wagmi-connector:
+    runs-on: ubuntu-latest
+    needs: publish-sdk
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: recursive
 
-      - name: Build wagmi connector package
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v4
+        with:
+          version: 9.11.0
+
+      - name: Use Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: lts/Iron
+          cache: 'pnpm'
+
+      - name: Install dependencies
+        run: pnpm install -r --frozen-lockfile
+
+      - name: Build the package
         run: pnpm nx build sdk
 
-      - name: Prepare wagmi connector package.json
-        working-directory: packages/sdk
-        run: node prepare-wagmi-connector.mjs
+      - name: Publish wagmi connector to NPM
+        run: node packages/sdk/publish-connector.mjs
         env:
           INPUT_VERSION: ${{ github.event.inputs.version }}
-
-      - name: Create .npmrc for NPM (wagmi connector)
-        run: echo "//registry.npmjs.org/:_authToken=${{ secrets.NPMJS_NPM_MATTERLABS_AUTOMATION_TOKEN }}" > ~/.npmrc
-      
-      - name: Publish wagmi connector to NPM
-        working-directory: packages/sdk
-        run: npm publish --access public
-    
-      - name: Create .npmrc for GitHub Packages (wagmi connector)
-        run: echo "@OWNER:registry=https://npm.pkg.github.com\n//npm.pkg.github.com/:_authToken=${{ secrets.GITHUB_TOKEN }}" > ~/.npmrc
-
-      - name: Publish wagmi connector to GitHub Packages
-        working-directory: packages/sdk
-        run: npm publish --access public
-
-      - name: Restore original package.json and README
-        working-directory: packages/sdk
-        run: |
-          mv package.json.backup package.json
-          mv README.md.backup README.md
+          NPM_TOKEN: ${{ secrets.NPMJS_NPM_MATTERLABS_AUTOMATION_TOKEN }}
